% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{covariance.function}
\alias{covariance.function}
\title{Covariance function}
\usage{
covariance.function(x1, x2, pars)
}
\arguments{
\item{x1}{a vector of random real numbers}

\item{x2}{a vector of random real numbers}

\item{pars}{a list containing hyperparameters and kernel specification}
}
\value{
the calculated kernel/covariance matrix of size (\code{length(x1)} x \code{length(x2)})
}
\description{
Calculates the covariance/kernel of a set of input points
}
\examples{
 x = rnorm(10)
 K = covariance.function(x, x, list(var=1, inv.scale=2, gamma=2, noise=0, kernel="gamma.exp"))
}

