% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{demo.regression}
\alias{demo.regression}
\title{Demonstration of GPR}
\usage{
demo.regression()
}
\description{
Run a GPR example. Generate samples using the prior GP, then learn the posterior GP and predict a new set of data points.
}

