% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{sample.from.gp}
\alias{sample.from.gp}
\title{Sample function values from a Gaussian process}
\usage{
sample.from.gp(x = seq(-10, 10, 0.25), m = NULL, K = NULL)
}
\arguments{
\item{x}{a vector of real values}

\item{m}{a vector of mean values, e.g. a vector of zeroes}

\item{K}{a symmetric positive semi-definite matrix of size \code{length(x)} x \code{length(x)}, e.g. calculated using \code{convariance.function}}
}
\value{
a vector of normal-distributed values
}
\description{
Sample function values from a Gaussian process
}
\examples{
 x = seq(-10, 10, .25)
 m = rep(0, length(x))
 K = covariance.function(x, x, list(var=1, inv.scale=2, gamma=2, noise=0, kernel="gamma.exp"))
 sample.from.gp(x, m, K)
}

